refactor: extract lookup/LDAP methods to TargetTypeLookup
