+++ attribute-manager/pom.xml
@@ -57,6 +57,11 @@
       <scope>runtime</scope>
     </dependency>
     <dependency>
+      <groupId>javax.xml.bind</groupId>
+      <artifactId>jaxb-api</artifactId>
+    </dependency>
+
+    <dependency>
       <artifactId>zm-common</artifactId>
       <groupId>zextras</groupId>
     </dependency>
