+++ soap/pom.xml
@@ -99,6 +99,10 @@
       <artifactId>jaxb-impl</artifactId>
       <scope>runtime</scope>
     </dependency>
+    <dependency>
+      <groupId>zextras</groupId>
+      <artifactId>attribute-manager</artifactId>
+    </dependency>
   </dependencies>
 
   <build>
