+++ soap/src/main/java/com/zimbra/soap/admin/message/GetDelegatedAdminConstraintsRequest.java
@@ -7,19 +7,16 @@ package com.zimbra.soap.admin.message;
 
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
-
+import com.zimbra.common.soap.AdminConstants;
+import com.zimbra.cs.account.accesscontrol.TargetType.SoapTargetType;
+import com.zimbra.soap.type.NamedElement;
 import java.util.Collections;
 import java.util.List;
-
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
-
-import com.zimbra.common.soap.AdminConstants;
-import com.zimbra.soap.type.TargetType;
-import com.zimbra.soap.type.NamedElement;
 
 /**
  * @zm-api-command-auth-required true
@@ -58,7 +55,7 @@ public class GetDelegatedAdminConstraintsRequest {
      * @zm-api-field-description Target Type
      */
     @XmlAttribute(name=AdminConstants.A_TYPE, required=true)
-    private final TargetType type;
+    private final SoapTargetType type;
 
     /**
      * @zm-api-field-description ID
@@ -87,7 +84,7 @@ public class GetDelegatedAdminConstraintsRequest {
     }
 
     public GetDelegatedAdminConstraintsRequest(
-                TargetType type, String id, String name) {
+                SoapTargetType type, String id, String name) {
         this.type = type;
         this.id = id;
         this.name = name;
@@ -105,7 +102,7 @@ public class GetDelegatedAdminConstraintsRequest {
         return this;
     }
 
-    public TargetType getType() { return type; }
+    public SoapTargetType getType() { return type; }
     public String getId() { return id; }
     public String getName() { return name; }
     public List<NamedElement> getAttrs() {
