+++ store/src/main/java/com/zimbra/cs/account/accesscontrol/ACLAccessManager.java
@@ -709,7 +709,7 @@ public class ACLAccessManager extends AccessManager implements AdminConsoleCapab
       boolean asAdmin)
       throws ServiceException {
 
-    TargetType targetType = TargetType.getTargetType(target);
+    TargetType targetType = TargetTypeLookup.getTargetType(target);
     if (!CheckRight.rightApplicableOnTargetType(targetType, rightNeeded, canDelegateNeeded)) {
       return false;
     }
