+++ store/src/main/java/com/zimbra/cs/account/accesscontrol/CheckAttrRight.java
@@ -59,7 +59,7 @@ public class CheckAttrRight extends CheckRight {
         super(target, rightNeeded, canDelegateNeeded);
 
         mGrantee = grantee;
-        mTargetType = TargetType.getTargetType(mTarget);
+        mTargetType = TargetTypeLookup.getTargetType(mTarget);
         mAttrRightNeeded = rightNeeded;
     }
 
@@ -95,7 +95,7 @@ public class CheckAttrRight extends CheckRight {
         // if the target is a domain-ed entry, get the domain of the target.
         // It is need for checking the cross domain right.
         //
-        Domain targetDomain = TargetType.getTargetDomain(mProv, mTarget);
+        Domain targetDomain = TargetTypeLookup.getTargetDomain(mProv, mTarget);
 
         if (!car.isAll()) {
             // check grants granted on entries from which the target entry can inherit
@@ -186,7 +186,7 @@ public class CheckAttrRight extends CheckRight {
 
         AllowedAttrs result;
 
-        AttributeClass klass = TargetType.getAttributeClass(mTarget);
+        AttributeClass klass = TargetTypeLookup.getAttributeClass(mTarget);
 
         if (car == CollectAttrsResult.ALLOW_ALL)
             result = processAllowAll(allowSome, denySome, klass);
